added menu and topbar from outside file

+++ SocialNetwork/SocialNetwork-war/src/java/beans/WallMB.java
@@ -110,7 +110,10 @@ public class WallMB {
         Integer fID;
         String idString = FacesContext.getCurrentInstance().getExternalContext().getRequestParameterMap().get("wallId");
         uID = (Integer)SessionUtils.getItem(SessionUtils.ID_KEY);
+        if(idString == null)
+            return false; 
         fID = Integer.parseInt(idString);
+
         return friendService.areFriends(uID, fID);
     }
     
@@ -120,11 +123,13 @@ public class WallMB {
             Integer fID;
             String idString = FacesContext.getCurrentInstance().getExternalContext().getRequestParameterMap().get("wallId");
             uID = (Integer)SessionUtils.getItem(SessionUtils.ID_KEY);
-            fID = Integer.parseInt(idString);
-            friendService.removeFriendship(uID, fID);
-            
-            ExternalContext context = FacesContext.getCurrentInstance().getExternalContext();
-            context.redirect(context.getRequestContextPath() + "/faces/wall.xhtml?wallId="+getUserId());
+            if(idString != null){
+                fID = Integer.parseInt(idString);
+                friendService.removeFriendship(uID, fID);
+
+                ExternalContext context = FacesContext.getCurrentInstance().getExternalContext();
+                context.redirect(context.getRequestContextPath() + "/faces/wall.xhtml?wallId="+getUserId());
+            }
         } catch (IOException ex) {
             Logger.getLogger(WallMB.class.getName()).log(Level.SEVERE, null, ex);
         }
